fixed 404 bug masquerading as 400 bug.
for reasons entirely unclear to me, when `in` is called, it is passed
the configured payload as well as a timestamp that always ends up with
"0001-01-01 00:00:00 +0000 UTC". This causes an index out of range error
in Azure, which is returned up to the user as an HTTP 400 error. if the
file doesn't exist, the user sees http 400 with "OutOfRangeInput".

this fix removes the snapshot functionality on Get() as there is not
enough of a strong reason to snapshot a file. if
needing a snapshot of a file is needed, there are much bigger
problems that can't easily be addressed by concourse.

now the error message will properly return with 404 instead of 400.

hours Mike wasted on this bug: 7.

Signed-off-by: Mike Lloyd <mlloyd@pivotal.io>