Fix out step not responding with version data
For blobs matched by regexes, the check step pulls out the version
number from the blob name and adds it to the response. This commit adds
the same behaviour to the out step so that Concourse sees versions for
the same blob produced by the out and check steps as being equal by
having both `path` and `version` fields with the same values.

Closes #44.